completed the MVP

+++ src/main/java/com/example/monoauction/security/SecurityUtils.java
@@ -19,6 +19,6 @@ public class SecurityUtils {
     }
 
     public static Long getCurrentUserId(){
-        return getCurrentUser().getId();
+        return (Long) getCurrentUser().getId();
     }
 }
